working serialize, searchItems
