+++ packages/search/src/ago/serialize.ts
@@ -1,6 +1,7 @@
 import { ISearchParams } from "./params";
 import { createFilters } from "./create-filters";
 import { paramSchema } from "./param-schema";
+import { encodeFilters } from "./helpers/encode-filters";
 
 /**
  * serialize raw query parameters into hub specific URI encoding
@@ -13,14 +14,12 @@ import { paramSchema } from "./param-schema";
  * @returns {string}
  */
 export function serialize(searchParams: ISearchParams): string {
-  // encode non-filtereable keys first
-  const nonFilterableKeys = Object.keys(searchParams).filter(
-    name => !isFilterable(name)
-  );
-  const encodedNonFilters = nonFilterableKeys
-    .map(key => `${key}=${encodeURIComponent(searchParams[key])}`)
+  // 1. handle simple params like non-filters like q, sort etc.
+  const encodedNonFilters = Object.keys(searchParams)
+    .filter(param => !isFilterable(param))
+    .map(param => `${param}=${encodeURIComponent(searchParams[param])}`)
     .join("&");
-  // in order to encode filters, first get a filters obj in terms of fn and terms
+  // 2. handle other params like tags, source, hasApi, groupIds that can be filtered
   const filters = createFilters(searchParams);
   const encodedFilters = encodeFilters(filters);
   return [encodedNonFilters, encodedFilters].join("&");
@@ -28,18 +27,4 @@ export function serialize(searchParams: ISearchParams): string {
 
 export function isFilterable(field: string) {
   return paramSchema[field] && paramSchema[field].type === "filter";
-}
-
-export function encodeFilters(filters: any = {}): string {
-  return Object.keys(filters)
-    .map(name => {
-      const attribute = filters[name];
-      const { fn, terms } = attribute;
-      // filters that are part of the catalog defenition are OR'd together then anded to the query
-      const type = attribute.catalogDefinition ? "catalog" : "filter";
-      const key = encodeURIComponent(`${type}[${name}]`);
-      const values = terms.map(encodeURIComponent).join(",");
-      return fn ? `${key}=${fn}(${values})` : `${key}=${values}`;
-    })
-    .join("&");
 }
